rtl8139: replace DEFINT_*() with spinlock
We must use a new RTL-driver spinlock, as upper layers may call rtl_tx() with
the NIC lock held.  Now that the lock exists, use it for protecting the
multicast data structures as well.

This does not manifest in a bug today, but could with preemption --- a thread
could be preempted in a critical section (as DEFINT_*() doesn't disable
preemption), and result in a race condition.