proc: don't run defints in POP_INTERRUPTS()
I remembered it's only the refcount test that causes a problem without this ---
otherwise tests pass fine.  So add a spin lock/unlock to that test to force
defints to run more frequently.  This may make the test somewhat less rigorous,
but saves several seconds in test runtime across the kthread and VFS tests
(others are largely unaffected).