ld: make LD SMP-safe
This protects everything except the TTY variable and sink with a spinlock.  I
don't think the TTY is ever set more than once.  The sink is only set once, at
construction, outside test code.  A future change should enforce that tty and
sink are only set at construction time.

The only functional change here is that we don't call the ld's sink with the
lock held.  This requires copying the ld's termios flags with the lock held,
then calling ld_term_putc() with the local copies of the flags.  This is
slightly awkward, but is the best I can come up with that works for both
callers, ld_write() and ld_provide().