memory: set heap size independently from heap-mapped region
This will allow us to constrain the size of the heap to smaller than the minimum
global mapping size.  I'll use this with TSAN to restrict the heap to be small
so we can pre-allocate all the shadow pages --- otherwise, we'd still have to
statically restrict the heap size, but it would be done implicitly rather than
explicitly.

Using virtual memory that's demand-allocated for the kernel heap is, TBH, a bad
idea to begin with, and I should change that.