tsan: add basic (incomplete) TSAN implementation
This is incomplete in many ways.  Adds some basic tests to validate
non-data-race behavior is not flagged as a race.  In the future will add tests
that catch data races as well.

The algorithm and design here is based on the public TSAN papers as well as
examining the LLVM/clang source code.  It is a more limited version of the
algorithm than clang supports (in particular, it doesn't support epoch resets,
atomics, etc, and is largely unoptimized).