malloc: conservatively align based on block size
To date, this has not caused a functional problem on any platforms.  However,
unaligned accesses are not guaranteed to work or be fast, in general.

This becomes an issue of correctness with TSAN, as the compiler assumes
alignment about particular data types that may not hold when allocated with
kmalloc(), and it may instrument unaligned accesses as aligned ones.

This may not completely solve the alignment issue --- I wouldn't be surprised if
there are rare and/or subtle alignment problems hiding in the codebase still.