tsan: add basic tsan-failure test
Due to the special handling of thread starts and synchronization to deal with
non-preemptible code in the kernel, we have to try hard to trigger this TSAN
race in the test.  There is probably a more robust way to do this --- I'll
figure that out for use in future tests, but locking this one in now as a basic
smoke test.