proc: allow querying for interrupt/defint/thread execution context
This logic is _slightly_ subtle so encapsulating in the kthread module.  Also,
add a test that defints can be interrupted, which it looks like we didn't have
before.

Note: I initially added assertions to kmutex and spinlock code that they weren't
running in inappropriate contexts, but they break due to page fault handling
(which is done in an interrupt context, but is allowed to block).  For the
future, consider adding tracking for when we're handling a page fault so we can
special-case that.