tsan: model interrupts and interrupt-safety
This is a bit hairy.  To model interrupts, we introduce a "virtual" thread for
the interrupt handlers, which is what we record (and check) accesses against,
rather than the thread that was itself interrupted.

The harder part is preventing over-synchronization via the interrupt thread.  In
reality, two APOS threads today will be well-synchronized if they use interrupt
disabling mechanisms, but that is not very interesting (we want to catch
thread-to-thread races even when they have interrupt critical sections in
between).

To do so, we limit the updates that can be published via the interrupt thread
--- only writes by the interrupt thread itself, not writes it has observed from
other threads, can be synchronized by synchronizing with the interrupt thread.

A lot of this complexity comes because we're mixing legacy code (which is
non-preemptible, and does in fact do non-trivial synchronization just with
interrupt critical sections) and new code, which we want to catch bugs in.  When
all code is upgraded and is preemptible, and all interrupt handler code is
SMP-safe (using spinlocks), this logic can be significantly simplified.