kthread: fix kthread_execution_context() defint bug
The code previously didn't distinguish properly between being in an
interrupt-triggered or thread-triggered defint.  Found this while adding defint
support to tsan, and noticing we'd fail to find interrupt/defint races when the
defint was triggered synchronously but the interrupt was interrupting the defint
itself (as those would both be considered the 'defint' thread, and therefore
non-racy).