tsan: fix shadow memory address calculation
This was not dividing by the memory cell size --- so it was taking a byte
offset (rounded down to a multiple of 8), and multiplying that by the shadow
cell size, causing a blowup of 8x of the resulting address.

This works fine until the heap usage starts to exceed 1/8 of the total heap, at
which point the shadow memory addresses calculated walk off the end of the
pre-allocated region.