tsan: ignore stack races with interrupt and defint threads
This is an unfortunate consequence of using TSAN to try and catch interrupt and
defint races.  Since interrupts and defints reuse the current thread's stack,
we will record accesses to stack addresses, which will look like races when a
normal thread uses the same stack addresses.

The actual synchronization on these addresses happens when functions enter/exit
and the stack is manipulated, but modeling a full synchronization with
interrupt/defint threads at those points would make it much harder to find
races.

There could be a more complicated but granular approach where we explicitly
synchronize only on stack addresses as threads enter/exit functions, but I'm not
sure that would buy much in practice.