memory: always mark kernel heap mappings as writable (and avoid double alloc)
This is an _old_ bug, probably as old as the original VM code.  The bug happens
when an anonymous kernel memory region (i.e. the heap) is accessed via a READ
before a WRITE.  The VM code will allocate a page for it, put a read-only
mapping in place, and continue.  The code that reads will read whatever happens
to be in memory (which is likely garbage, or zeroes).  A subsequent WRITE to
that address will trigger another page fault, which will hit this branch and
cause a _different_ page to be allocated and new mapping created.

There are two problems with this,
a) no well-formed code should be reading fresh memory, as it is garbage
b) if (a) occurs, we will waste an allocated page

(a) happens in practice on riscv64 with GCC, though I never noticed until commit
a2184194e01352d ("riscv64: assert that page mappings don't change"), which I
unfortunately didn't test with GCC before committing.

Here is a snippet of init_block() (from kmalloc.c) that GCC generates
illustrating the problem (unnecessary loads before writes):
 #   b->magic = KALLOC_MAGIC;
       42:       fd843783                ld      a5,-40(s0)
       46:       fab00713                li      a4,-85
       4a:       00e78023                sb      a4,0(a5)
 #   b->free = true;
       4e:       fd843783                ld      a5,-40(s0)
       52:       4705                    li      a4,1
       54:       00e780a3                sb      a4,1(a5)
 #   b->length = 0;
      58:       fd843783                ld      a5,-40(s0)
      5c:       0047c703                lbu     a4,4(a5)
      60:       8b01                    andi    a4,a4,0
      62:       00e78223                sb      a4,4(a5)

For b->magic and b->free, a new immediate is stored blindly (correct).  For
b->length, for some reason, GCC loads the current value, bitwise-ands it with
the literal zero, then stores that back.

To fix this, we simply treat all accesses to anonymous kernel areas as writes.
There is no COW optimization that will be done, so even if we tracked the
allocated page and could upgrade it from R/O to R/W, we'd be incurring an extra
fault for absolutely no reason.