tsan: model spinlocks as TSAN locks
Had to play around with the tests to make them work --- both the creating thread
and the other test thread need to be preemptible for the test to work properly.
This is another good argument for making premptibility a thread-creation-time
flag, as it would let us avoid the explicit synchronization point in the thread
trampoline for preemptible threads.