memory: convert kmalloc to spinlock
This requires slight special handling, as kmalloc() can be called before kthread
is fully initialized --- which means that spinlocks can't be used.  We get
around that by checking if the kthread module is initialized and if not, simply
disabling interrupts.  This assumes that before kthread is initialized, there is
no risk of SMP.