vfs: convert poll to use spinlock
The simplest thing is a global poll spinlock, which mirrors the current behavior
(blocking interrupts).  More fine-grained locking is much more complex, as it
would need to deal with two separate lifetimes (the lifetime of the poll call,
and the lifetime of the object being polled), and locking from two different
directions into the shared poll_ref_t state.