devicetree: fix unaligned 64-bit access with utility
The compiler assumes, when doing a naked cast of the pointer to uint64_t, that
the address is properly aligned for a uint64_t.  On currently-supported
architectures, I don't think this creates a correctness problem, but it may
cause a performance issue.

This becomes an actual problem when compiled with TSAN, when the alignment of
the access becomes important to which TSAN hook is called.

I've added a utility header for wrapping unaligned loads/stores, such that the
implementation can be changed based on the underlying platform.  It currently
just uses kmemcpy.