futex: switch to mutex
This requires a slight fiddling to do the module lazy initialization, since
mutexes can't be statically initialized.  This also makes the futex code
preemptible (which it was not previously --- it was racy if preempted in between
the spinlock unlock and the sleep), and removes a non-locked
scheduler_wait_on_*() call.