block cache: wait on g_mu in flush thread
This is not much of a functional change --- the only difference is that we don't
unlock and relock the block cache mutex after each flush.  However, in
flush_cache_entry(), the mutex is unlocked/relocked anyway around the actual
underlying write operation.

This removes a call to scheduler_wait_on_interruptable(), and fixes a race
condition on g_flush_queue_wakeup_queue, which is not currently protected
properly even though this thread is marked as preemptible.