cbfs: fix getdents skipping-entries bug
The bug is caused by double-counting the skipped entries when calculating the
offset of generated dirents.  This bug dates back to 2014 commit
7e6673177 ("cbfs: implement dynamic directories").  An existing test actually
catches it (on 64-bit platforms at least), but was written incorrectly and so
masks the bug.

As of commit aaceb5f6318c ("vfs: ensure getdents always generates aligned
kdirent_t's"), the standard tests fail as well on i586, but I didn't notice at
the time.