ata: refactor ATA code and make SMP-safe/preemptible
This is messy old code that I don't particularly like.  This cleans it up a
little bit, but not much --- there's still weird handoff of scheduling
responsibilities between ata.c and dma.c, for example.

This makes the code preemptible and interruptable.  Not all the interruptible
paths are tested.  I've stared at the new code a fair amount, but have not
tested all the code paths, so it's possible there are still bugs.  This whole
module probably deserves a rewrite (and that would have been less time in the
end than all this surgery I've been banging my head against).

I could simply delete the ATA driver and use NVMe all the time, but I think this
still would be required to boot on my old test hardware, which I'd like to try
again.

This removes the last non-kthread/scheduler use of scheduler_wait_on() --- so
all IO code is now interruptable.