tsan: fix extended-entry tsan_find_access() bug
If the is_read bit didn't match, we would skip the entry immediately, then try
to interpret the next entry (which is the extension of the current one) as its
own access.

This could also be avoided by dedicating a bit to whether an entry is an
extension or not.