tsan: release to interrupt thread before each thread switch
This in particular prevents a problem after thread exit --- an exiting thread
will never have a chance to release its final accesses to the interrupt thread,
which will cause spurious races to be seen later when the memory is reused then
freed from an interrupt context.  This generally triggers from the timer
interrupt.

Rather than special-casing this in proc_thread_exit(), I'm doing it in
kthread_switch(), which means we'll pick up all accesses through the scheduler
and kthread code as well.