tsan: prevent recursive tsan calls in fnv_hash_array_continue()
tsan_check_internal() calls fnv_hash_array() when a shadow word is evicted which
in turn (now) calls tsan_check_range().

This also adds a test that explicitly exercises the overwrite-shadow-cell case.
It does not reliably trigger this bug, though --- for that, there must be an
overwrite on the thread stack itself to cause the infinite recursion.  It's
probably possible to write a test for that, but difficult.