tsan: treat PUSH_AND_DISABLE_INTERRUPTS() as cross-thread sync
This models, in TSAN, the actual behavior of the system, which is that
interrupt-disabling is (currently) a full synchronization between different
threads, not just between a thread and an interrupt handler.

This means that, when enabled system-wide, TSAN won't be able to catch ~any
races, as effectively all code relies on underlying thread or scheduler
primitives that are still based on PUSH_AND_DISABLE_INTERRUPTS() rather than
spinlocks.

That will be OK, though, as it means we'll catch races as they're actually
introduced (as the code is converted), rather than trying to catch code that
_will_ be racy in the future when preemptible/SMP-safe.  The latter can still be
done by disabling the legacy-full-sync mode, with the caveat that TSAN will then
flag a lot of races incorrectly in the kthread/scheduler code that hasn't been
updated yet.