tsan: fix kmutex unlock release ordering
Previously, the release happened _after_ a yield to a waiting thread.  That
means that a fale positive race could be triggered if the waiting thread
accessed the same memory as the thread that is releasing the lock.

None of the existing tests exercise this case, so added a new one to do so.  It
has to explicitly yield in the critical section to ensure the other thread is
waiting on the lock.