tsan: do full legacy interrupt sync around scheduler yield
This is required to release any changes made since PUSH_AND_DISABLE_INTERRUPTS()
was called if POP_INTERRUPTS() wasn't called before we yield.  This triggers a
(false positive) race in the full kernel TSAN mode on accesses in the
kthread/scheduler code.

I wasn't able to write a test that demonstrates why the release to the
interrupt-thread is also needed, but that's the first form of this I found in
the full kernel code.

Once the kernel is fully switched to spinlocks and PUSH_AND_DISABLE_INTERRUPTS()
isn't used for cross-thread synchronization, all this can be deleted.