tsan: fix release to interrupt thread before yielding
The previous commits made the release to the interrupt thread (the limited
version) conditional on whether the legacy full-sync mode is enabled, which is
incorrect.  Only the legacy full sync (to the fake interrupt lock --- i.e.
the sync that synchronizes _across_ threads, not just with interrupt handlers)
should be conditional.

Adds a test that catches this explicitly by doing PUSH_AND_DISABLE_INTERRUPTS()
then an access.