tsan: have legacy_interrupt_sync be per-thread rather than global
This way, when the TSAN tests run they can only disable it for themselves rather
than for all threads.

Unclear if this will actually help with the current problem I have (which is
running the TSAN tests with the required changes to kmalloc's fill_buffer  in
place).  With this change, the main thread now won't be affected by the TSAN
test threads' settings, but if the race is with the TSAN test thread itself,
then we'll still have a problem.