tsan: monitor for races in kernel data (.data/.bss)
This requires changing the check for whether the address is instrumented to be
dynamic rather than static, as the bounds of .bss/.data isn't statically known.
It _is_ known at link-time, however, so the check is able to be done without
loading any memory.

We also have to deal with more complex shadow-mapping setup code.  We now set up
multiple mappings (one for the heap, and one for the .data/.bss section).  Each
of these gets two allocations, one for the shadow region, and one for the page
metadata region.  The page metadata code relies on the regions being mapped not
being too close (or it will try and double-allocate page metadata pages causing
a panic).

Other than setup and range checking, the rest of the code works the exact same.