riscv: add software-interrupt based preemption #tsan
This resolves a data race that currently happens in sched_preempt_me() when it
calls sched_tick(), which accesses g_run_queue without locking.

Adding interrupt protection around the check of g_run_queue would fix the
problem, but also introduce explicit cross-thread synchronization at every
preemption hook point taken, which would largely defeat the purpose of it.

One way around this would be to add special handling to sched_tick() to make it
seem as if it were running in an interrupt context (which it effectively should
be, as it is synchronously imitating an interrupt).  That is clumsy and
difficult.

Instead, this adds a supervisor-software driven interrupt mechanism for riscv64.
This allows sched_preempt_me() to synchronously trigger an actual interrupt,
which will be dealt with the same as any other interrupt (i.e. exactly the
desired behavior).

On other architectures (where we don't support TSAN anyway), wraps the call to
sched_tick() in interrupt guards to solve the race condition.

This was found in a test run under whole-kernel TSAN.