tsan: switch shadow cell is_write -> is_read
This makes it consistent across types.  is_write seemed more natural at first
(as writes are "more interesting" than reads), but is_read is actually the more
important bit --- accesses are safe if both are reads, which means we can
express "is safe" as an AND operation rather than a NOR operation.  This also
matches the polarity of is_atomic (both accesses are safe if they're both
atomics or both reads).

Not that we take advantage of this, but a vectorized version of this like clang
has could now easily express this as AND'ing the access bits between two shadow
cells, and if it's nonzero (both reads OR both atomics), it's safe.