atomics: align/pad atomic32_t to 8-byte boundaries
In writing tests for TSAN atomic synchronization, I found that there was often
false sharing between the atomic in the test and nearby data which caused
spurious false negatives due to shadow cell eviction.  Given that atomics are
likely often used for synchronization, and are likely stored near data they are
synchronizing with, this padding/alignment ensures the atomic always gets a
dedicated shadow memory cell.