atomics: implement ATOMIC_SEQ_CST
On the TSAN side, this treats ATOMIC_SEQ_CST the exact same as ATOMIC_ACQ_REL,
which AFAICT is the same thing clang's TSAN implementation does.  I'm not sure
how to model the difference between them re the global memory order (or if that
even can be modeled in TSAN).