proc: use atomic for kthread_t::preemption_disables
This value is only ever read from the current thread context or from an
interrupt that has interrupted the current thread.  Therefore we don't need any
ordering or synchronization, just atomicity of accesses.

Actually using the atomics API really reinforces how verbose it is.  I also see
the appeal better of Linux-style READ_ONCE/WRITE_ONCE --- there are a number of
places where we access this when interrupts are already disabled, so we could do
a plain access (which should compile to the same thing, but is much less
verbose).