proc: add reference count to process_t
This solves the lifetime issue of process_t before we can make it fully
locked/SMP/preemptible-safe.  It was surprisingly painless --- most code that
calls proc_get() is test code, and I'm just not bothering to update that (it
doesn't need to be resilient to concurrent process exit the way user-triggered
code does).

There are far more users of proc_current() than proc_get() (almost 900!), but
there's no need to refcount there --- the current process is (by nature of being
currently run) guaranteed to survive until the current thread is done with the
reference, definitionally.