tcp: fix data race on socket->state when sending datafin
socket->state is currently read without the lock held.  This was caught with
static analysis, not TSAN (possible TODO --- when TSAN support is working better
for full-kernel, see if I can catch this race with it).

This is a bit of a bandaid --- in theory this should be unnecessary, since
tcp_send_datafin() immediately relocks the socket and checks the state again.
However there is a brittleness currently where if we get into tcp_send_datafin()
from this path in SYN_RCVD, then we'll trigger an assertion failure on the
ACK-received-before-SYN/ACK-sent race case (a rare one, indicating incredibly
bad luck or a malicious peer).

This commit keeps the current (benign) racey behavior but leaves a fat TODO to
fix this in a better way.

Committing this separately from the changes adding thread safety annotations to
the TCP code, as this is a functional change.