proc: add optional kmutex_t thread safety annotations
The thread safety annotations are incompatible with KMUTEX_AUTO_LOCK, so I need
to remove that.  There is a _lot_ of code that uses it, however, and the thread
safety annotations are themselves very useful for doing that change (as it can
find code paths that don't unlock properly).

This approach lets us opt-in specific modules while I work on them to remove
KMUTEX_AUTO_LOCK.  Once it is removed everywhere, I can enable these
unconditionally, and also start using the kmutex thread annotations.