vfs: remove KMUTEX_AUTO_LOCK from core VFS code
This fixes at least one place where, in theory, we could have unlocked the mutex
after the file or vnode was deleted.

The core VFS code gets pretty wonky with how it does locking dynamically, so
this will require a bunch of cleanup work to play nice with thread safety
annotations when they're fully enabled.