proc: add GUARDED_BY to process_t::fds
Looking at proc_finish_exit(), this will likely need a larger restructuring as I
protect more of process_t with the mutex, since this is currently a mix of
direct manipulation of the process's fields and indirect calls (such as
calling proc_alarm_ms() or proc_force_signal_group()).

The funkiness with the FD table in exit() and exec() is a sign of things to come
--- the layering in the code is not well established and it's highly reentrant
between process code and VFS code (which is bad).  Will need to untangle that as
part of cleaning up these locks.