thread-safety: fix annotations to work with -Wthread-safety-pointer
-Wthread-safety-pointer is a proposed addition to clang to catch cases when
pointers to guarded objects are taken (which makes it much more effective for C
codebases).

I've compiled a custom clang and these changes are required to make the static
analysis happy.  There are no functional changes other than to change the unlock
path to protect nic's IPv6 iface_id/iface_id_len, which are today in practice
const-after-construction, but are not guaranteed to be in theory (so this
locking is, while annoying, correct).