proc: split setpgid() into policy and mechanics
This is similar to the signal code, where there is a kernel-internal version
that forces a signal onto a process or thread.

This split makes setsid() simpler, and will make things much more natural when I
add locking.  I expect to do similar changes to many other parts of the proc
code to make locking of different types easier.