proc: add defint timers
Most timer users actually want their function to run in a defint rather than
interrupt context --- and they currently each bounce it through a custom defint.
This is a common mechanism that (a) is SMP-safe, and (b) removes duplication.

The API is a little sharp, as it doesn't give a way for the caller to ensure a
timer has run or been cancelled, so the caller must do this on their own.  For
refcounted objects (such as a TCP socket or process_t) this shouldn't add any
overhead to the existing lifetime management.  Doing it this way avoids blocking
calls and therefore deadlock possibility