proc: switch alarm() to use defint timer
The behavior here is now slightly less well-defined --- previously, in a non-SMP
context, the call to proc_alarm_ms() was guaranteed to cancel the existing
alarm.  It was not well-defined in an SMP context.

Now, it's possible for the cancellation and the running of the timer to race.  I
think that's still POSIX-compliant (which doesn't require alarm() to block if a
SIGALRM is being processed), and better-defined in an SMP environment anyway.
If the call and alarm race, it is the same as if there is a delay in signal
delivery or processing by the application.