ld: refactor SIGTTIN code into tty_check_read() helper
This is not useful in the same way tty_check_write() is, as it is only used
once.  However it puts the SIGTTIN code and SIGTTOU code together in one module
(tty_util.c), and gets the TTY code out of the ld module.  This code will also
get (slightly) more complex soon with process locking.