proc: fix bug in wait() if process's pgroup changes
Currently we read the process's current process group once at the start of
waitpid().  However, if the process group changes while waitpid() is blocking,
then it will return a child from the _old_ process group, not the new (current).

This changes waitpid() to re-read the process group each time through the wait
loop, so (if pid == 0) it will only find children with the same _current_
process group.

This also beefs up the existing waitpid() test slightly, by (a) adding a fourth
child whose process group does not match, and (b) ensuring that we aren't
executing the test with pgroup == 1 (which creates ambiguity in waitpid() when
-1 is passed --- it is interpreted [correctly] as "any child" rather than
"children in my current pgroup [which happens to be 1]".