proc: snapshot process group before sending signals
This is clumsy and inefficient, but I think should be more than sufficient for
now.  This is semantically similar to what Linux does, but Linux uses an RCU
list rather than doing a fully copy (under lock!) of the current process group
contents.

As of this commit it is pure overhead, but this will allow us to snapshot the
process group contents with the process table lock held (when that is added),
THEN send signals to each process without the process table lock, avoiding
complex lock ordering issues.