proc: add global process table/process group/session lock
This lock globally protects the contents of the process table (i.e. the process
ID to process mapping, though not the processes themselves), process group
membership, and session data.  All of those should change infrequently enough
that a global lock is fine.

In most places this is a mechanical change, though it required some refactors
(see preceding commits).  Signal handling (assigning signals to process groups)
remains an issue, particularly in tcgroup functions --- future changes will
change how that works to make it safe even with locks protecting signal handlers
and thread lists.

This breaks the TSAN tests that verify races with dead threads (presumably
because this introduces synchronization on g_proc_table_lock, suppressing the
races the tests are trying to induce).  I'll work on fixing that next.