proc: don't lock in proc_set_current()
This is (currently) only needed to assert that the process table is valid, and
there is no data race possible so long as it is.  Locking here introduces
synchronization between threads whenever proc_set_current() is called, which is
currently on every thread switch.

This, in turn, breaks the TSAN tests (in particular the dead-thread tests),
since unwanted synchronization between the threads is added, suppressing the
race.  I'm actually not sure why it doesn't break the other TSAN race tests.

This will cause a problem again in the future depending on how we handle locking
for g_current_proc.  That will need to be a CPU-local variable so maybe we
manage it through that mechanism.

This represents a limited case of a more general problem --- there will be a
_lot_ of synchronization between threads in core code like the scheduler.  In
the future it may be necessary to split TSAN into two modes: scheduler-TSAN will
instrument scheduler code and be able to catch races there at the cost of
over-synchronizing other code (and suppressing races).  Non-scheduler-TSAN will
not instrument scheduler code (or locks), and in turn be much more likely to
catch races in other parts of the kernel.