proc: add process spinlock for signal state and thread list
As expected, signal handling is by far the most complex part of this.  This is,
in the end, a mostly mechanical change with some of the helper changes already
made.

I decided to use a single process-wide spinlock to protect all thread signal
state, rather than a per-thread spinlock, to make it possible to get a single
atomic view of the process's thread state.  This is required for e.g. the TTY
checks that need to determine if SIGTTIN/SIGTTOU is deliverable to _any_ thread
in the process.

Unfortunately there is no way (AFAICT) to express cross-object locking
requirements --- if I add GUARDED_BY(&process->spin_mu) to the kthread_data_t
members, the locking is never recognized.  I may try and come up with a way
around this (e.g. with a per-thread token capability that represents the process
lock), but in the meantime, those annotations are not possible.