proc: add analysis-only process_spin_mu to kthread_t
This is a bit of an experiment --- because clang's thread-safety analysis can't
follow pointers between objects, this adds a "fake" (no runtime cost) capability
to each kthread_t that shadows the parent process's spin_mu.  We then
lock/unlock the parent's spin_mu via a helper that locks both the actual lock,
and the fake capability.

This lets us annotate the thread's signal data with GUARDED_BY(process_spin_mu).

The only runtime cost here is the additional asserts in the for-all-threads
loops in signal.c, but those could be easily elided if necessary.

I _think_ this is an overall readability and safety win.