proc: protect {r,e,s}{uid,gid} with g_proc_table_lock
To determine if a signal can be sent from process A to process B, we have to
read these fields on both processes.  Rather than deal with the lock ordering
issues that introduces (as we otherwise assume we will always lock parent before
child), protect these fields with the global process table lock.  They should be
changed infrequently, so the contention introduced should be minimal.