vfs: add vfs_close_locked()/vfs_close_file()
This will make it easier to close the open file descriptors in proc
exit/exec/etc.

The TCP tests have a hook that causes the process lock to be reentrantly locked
(which panics the kernel) with this change, since the final file_unref() is now
called with the process lock held.  This changes the TCP test to not trigger
that path.