proc: handle setpgid() called on zombies
This is a pre-existing race condition.  We remove the process from the process
group list in proc_finish_exit(), but don't update pgroup (because the parent
needs to read it in waitpid()) --- this means anyone expecting them to be
consistent is in for a surprise.

Rather than having everyone treat zombies differently, just wait to remove the
process from the pgroup until cleanup happens in waitpid().

This adds another lock of g_proc_table_lock in waitpid().  I think it's
logically harmless, but should be fixed with consistent locking in waitpid().