proc: add SWAIT_NO_SIGNAL_CHECK scheduler_wait() flag
This will be used before returning to user-space when checking if the current
process/thread is stopped.  This is currently a no-op, but once we add locking
to protect the process state (which we will then scheduler_wait() on), we need
to ensure that scheduler_wait() won't check signals (which requires recursively
taking the process lock).

It's a bit inelegant, but I can't think of a better approach off the top of my
head.