proc: protect process->state with spin_mu
This is a largely uninteresting change other than in wait and user_prepare.  In
proc_prep_user_return(), there are two (logical) race conditions that will
manifest when preemption is enabled in this code, related to signal handling.

In finish_waitpid(), we have to do a bit of a lock dance because we can't unlock
spinlocks in a different order than they were acquired.  We also must slightly
change the teardown order so that we don't call vm_area_destroy() (which can
block) with the spinlock held.