proc: protect process exit_status and exiting with spin_mu
exit_status is only accessed from a user context, so theoretically could use mu
rather than spin_mu, but needs to be examined in the inner loop of
proc_waitpid() while g_proc_table_lock is held.

exiting could be protected with mu, but is already only accessed in places where
we need to take the spinlock anyway, so protect it with spin_mu for simplicity.