proc: handle spinlocks held over preemptible thread creation
Currently if you're holding a spinlock while creating a child thread, it will
not be marked preemptible even if the parent is overall.  This takes the number
of spinlocks held into account to propagate preemptibility in that scenario.
This will be required when we lock a spinlock in proc_thread_create() around the
call to kthread_create().

This was caught by an assertion in the TSAN tests, so adding an explicit kthread
test for it as well.