memory: add heap profiling for raw page allocs
Currently these are untracked, and I know there is a leak (as if I run tests
over and over again in a loop, we eventually run out of memory).

This tracks allocation stack traces for all page frame allocations, and logs
them as part of the heap profile dumped by kmalloc with the `hp` command.  This
will cause any page frame leaks to show up as heap leaks in the heap profile.

This also fixes some incorrect assertions (which were comparing stack_idx, an
index of entries, against stack_size, which is a byte count).