riscv: set global bit on global mappings
This requires global mappings to be global from the top-level down --- i.e.
global mappings must be "strong" global mappings (guaranteed shared between all
processes, with the intermediate page tables also always shared).  It would be
possible to support "soft" global mappings as well (the mappings themselves are
marked global, but the page tables are not shared), but that currently wouldn't
be utilized.

This updates rsv_get_pte() to set flags on entries created all the way down.
Currently the only flag supported is RSV_PTE_GLOBAL.  I'm not sure if
RSV_PTE_USER makes sense (or is supported), but as we currently only use
hugepages for static kernel mappings, it's not important.

This will be required for the next change, which frees intermediate page tables,
but uses the global bit to ignore shared intermediate tables.