riscv: fix register allocation in rsv64_sbi_call()
Previously this let the compiler allocation registers and didn't specify that it
clobbers a0/a1/a6/a7.  If the compiler decided to store any of the local
variables in those registers, then there would be a conflict and the wrong
values would get set.  This happens if you compile with -O1.

Specifying clobber registers fixes the bug, but also then introduces additional
unnecessary moves.  This version specifically places the values in the necessary
registers and lets the compiler set up the moves.

Comparing the generated inner part of the function (all with -O1):
Without this change:
  mv  a0,a3
  mv  a1,a4
  mv  a7,a0  # bug!
  mv  a6,a1  # bug!
  ecall
  mv  a0,a0
  mv  a1,a1
  beqz  a2,ffffffff803c2038 <rsv64_sbi_call+0x1c>
  sd  a1,0(a2)

With clobber registers:
  mv  t0,a1  # unnecessary :(
  mv  a5,a0  # unnecessary :(
  mv  a0,a3
  mv  a1,a4
  mv  a7,a5
  mv  a6,t0
  ecall
  mv  a3,a0
  mv  a4,a1
  beqz  a2,ffffffff803c203c <rsv64_sbi_call+0x20>
  sd  a4,0(a2)
  mv  a0,a3  # unnecessary :(

With this change:
  mv  a7,a0  # compiler orders these correctly :)
  mv  a6,a1
  mv  a0,a3
  mv  a1,a4
  ecall
  beqz  a2,ffffffff803c2034 <rsv64_sbi_call+0x18>
  sd  a1,0(a2)